Rename generate previews command
