+++ harvester/core/management/commands/generate_previews.py
