+++ harvester/harvester/tasks/background.py
@@ -45,7 +45,7 @@ def harvest(seeds_source=None, reset=False):
         # Aggregating the metadata and content into the dataset
         call_command("update_dataset", f"--dataset={dataset.name}", "--no-progress")
 
-        call_command("generate_browser_previews", f"--dataset={dataset.name}", "--no-progress")
+        call_command("generate_previews", f"--dataset={dataset.name}", "--no-progress")
 
         # Based on the dataset we push to Elastic Search
         extra_push_index_args = ["--recreate"] if reset else []
